+++ options.js
@@ -151,7 +151,6 @@
         const username = document.getElementById('username').value;
         const apiToken = document.getElementById('password').value;
         const baseUrl = baseUrlInput.value;
-        const jql = document.getElementById('jql').value;
         const experimentalFeatures = experimentalFeaturesToggle.checked;
         const issueDetectionEnabled = issueDetectionToggle.checked;
         const frequentWorklogDescription1 = document.getElementById('frequentWorklogDescription1').value;
@@ -164,7 +163,6 @@
             username,
             apiToken,
             baseUrl,
-            jql,
             experimentalFeatures,
             issueDetectionEnabled,
             frequentWorklogDescription1,
@@ -202,7 +200,6 @@
           username: '',
           apiToken: '',
           baseUrl: '',
-          jql: '(assignee=currentUser() OR worklogAuthor=currentUser()) AND status NOT IN (Closed, Done)',
           experimentalFeatures: false,
           issueDetectionEnabled: true,
           frequentWorklogDescription1: '',
@@ -215,7 +212,6 @@
           document.getElementById('username').value = items.username;
           document.getElementById('password').value = items.apiToken;
           baseUrlInput.value = items.baseUrl;
-          document.getElementById('jql').value = items.jql;
           experimentalFeaturesToggle.checked = items.experimentalFeatures;
           issueDetectionToggle.checked = items.issueDetectionEnabled !== false;
           document.getElementById('frequentWorklogDescription1').value = items.frequentWorklogDescription1;
@@ -228,10 +224,6 @@
           sidePanelToggle.checked = items.experimentalFeatures && items.sidePanelEnabled;
 
           jiraTypeSelect.dispatchEvent(new Event('change'));
-
-          const apiExtension = items.jiraType === 'cloud' ? '/rest/api/3' : '/rest/api/2';
-          const jira = await JiraAPI(items.jiraType, items.baseUrl, items.username, items.apiToken);
-          const issues = await jira.getIssues(0, items.jql);
         });
     }
 })();